:recycle: move snowflake_time function to public.py
